perf(#1579): D3 — stream the multi-generation merge (O(window), not O(table))
Extract cross-generation read reconciliation into a new generation_merge.rs
submodule (issue #1116 campsite split; mod.rs -356 net lines) and add a
streaming driver:

- stream_generations_for_read: drives the KWayMerger on a blocking task and
  blocking_sends each stepped partition's live rows straight into the bounded
  channel. scan_stream's multi-generation branch now streams instead of
  collect+sort+dribble — live heap O(one partition + channel), time-to-first-row
  O(first partition). Merger-construction errors are signalled over a oneshot so
  scan_stream still falls back to the lazy per-reader merge.
- Emission order is the merger's (token,key) order = byte-identical to the
  collect+sort path's sort_by_token_order (pinned by the order test).
- merge_generations_for_read{,_with_metadata} gain a target_key param: point
  reads (scan_partition_clustering, scan_partition_with_cell_metadata multi-
  candidate) now keep ONLY the target partition and stop, replacing full-merge-
  then-retain; the metadata TTL scan is bounded to the target too. Byte-identical
  to the former retain(key==partition_key).
- Shared partition_live_rows / push_metadata_rows kernels so tombstone filtering
  is not forked across the three drivers.

Tests: memory guard now GREEN (peak resident 1 vs 1 across 40x fixture, was
34 vs 1371); order + 957 + 1578 parity all green.
